Processed review comments. Added some additional fixes.
