+++ addons/binding/org.openhab.binding.denonmarantz/OSGI-INF/.gitignore
@@ -1,0 +1,1 @@
+*.xml
