--- addons/binding/org.openhab.binding.denonmarantz/OSGI-INF/DenonMarantzHandlerFactory.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2014-2017 by the respective copyright holders.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
--->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="binding.denonmarantz">
-
-   <implementation class="org.openhab.binding.denonmarantz.internal.DenonMarantzHandlerFactory"/>
-
-   <service>
-      <provide interface="org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory"/>
-   </service>
-
-</scr:component>
