+++ addons/binding/org.openhab.binding.denonmarantz/about.html
@@ -8,7 +8,7 @@
 <body lang="EN-US">
 <h2>About This Content</h2>
  
-<p>&lt;<em>May 4, 2015</em>&gt;</p> 
+<p>November 19, 2017</p> 
 <h3>License</h3>
 
 <p>The openHAB community makes available all content in this plug-in (&quot;Content&quot;).  Unless otherwise 
@@ -23,17 +23,5 @@ apply to your use of any object code in the Content.  Check the Redistributor's 
 provided with the Content.  If no such license exists, contact the Redistributor.  Unless otherwise
 indicated below, the terms and conditions of the EPL still apply to any source code in the Content
 and such source code may be obtained at <a href="http://www.openhab.org/">openhab.org</a>.</p>
-
-        
-        <h3>Third Party Content</h3>
-        <p>The Content includes items that have been sourced from third parties as set out below. If you 
-        did not receive this Content directly from the openHAB community, the following is provided 
-        for informational purposes only, and you should look to the Redistributor's license for 
-        terms and conditions of use.</p>
-        <p><em>
-        <strong>FasterXML 2.3.2</strong> <br/><br/>
-        <a href="http://wiki.fasterxml.com/JacksonHome">Jackson JSON Processor</a> obtained from Maven central repository under 
-        <a href="http://www.apache.org/licenses/LICENSE-2.0.txt">Apache License</a>.
-        </em></p>
 </body>
 </html>
