+++ addons/binding/org.openhab.binding.denonmarantz/src/main/java/org/openhab/binding/denonmarantz/internal/DenonMarantzHandlerFactory.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2010-2017 by the respective copyright holders.
+ * Copyright (c) 2010-2018 by the respective copyright holders.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -17,7 +17,9 @@ import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
+import org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory;
 import org.openhab.binding.denonmarantz.handler.DenonMarantzHandler;
+import org.osgi.service.component.annotations.Component;
 
 /**
  * The {@link DenonMarantzHandlerFactory} is responsible for creating things and thing
@@ -25,6 +27,7 @@ import org.openhab.binding.denonmarantz.handler.DenonMarantzHandler;
  *
  * @author Jan-Willem Veldhuis - Initial contribution
  */
+@Component(service = ThingHandlerFactory.class, immediate = true)
 public class DenonMarantzHandlerFactory extends BaseThingHandlerFactory {
 
     private final static Set<ThingTypeUID> SUPPORTED_THING_TYPES_UIDS = Collections.singleton(THING_TYPE_AVR);
