+++ addons/binding/org.openhab.binding.denonmarantz/src/main/java/org/openhab/binding/denonmarantz/internal/DenonMarantzState.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2010-2017 by the respective copyright holders.
+ * Copyright (c) 2010-2018 by the respective copyright holders.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -56,6 +56,10 @@ public class DenonMarantzState {
 
     public DenonMarantzState(DenonMarantzStateChangedListener listener) {
         this.listener = listener;
+    }
+
+    public void connectionError(String errorMessage) {
+        listener.connectionError(errorMessage);
     }
 
     public State getStateForChannelID(String channelID) {
