+++ addons/binding/org.openhab.binding.denonmarantz/src/main/java/org/openhab/binding/denonmarantz/internal/connector/http/adapters/OnOffAdapter.java
@@ -1,12 +1,12 @@
 /**
- * Copyright (c) 2010-2016 by the respective copyright holders.
+ * Copyright (c) 2010-2018 by the respective copyright holders.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.denonmarantz.internal.communication.adapters;
+package org.openhab.binding.denonmarantz.internal.connector.http.adapters;
 
 import javax.xml.bind.annotation.adapters.XmlAdapter;
 
@@ -20,7 +20,7 @@ public class OnOffAdapter extends XmlAdapter<String, Boolean> {
     @Override
     public Boolean unmarshal(String v) throws Exception {
         if (v != null) {
-            return v.toLowerCase().equals("on");
+            return Boolean.valueOf(v.toLowerCase().equals("on"));
         }
 
         return Boolean.FALSE;
