+++ addons/binding/org.openhab.binding.denonmarantz/src/main/java/org/openhab/binding/denonmarantz/internal/connector/http/adapters/StringAdapter.java
@@ -1,12 +1,12 @@
 /**
- * Copyright (c) 2010-2016 by the respective copyright holders.
+ * Copyright (c) 2010-2018 by the respective copyright holders.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.denonmarantz.internal.communication.adapters;
+package org.openhab.binding.denonmarantz.internal.connector.http.adapters;
 
 import javax.xml.bind.annotation.adapters.XmlAdapter;
 
@@ -21,11 +21,11 @@ public class StringAdapter extends XmlAdapter<String, String> {
 
     @Override
     public String unmarshal(String v) throws Exception {
-        if (v != null) {
-            v = StringUtils.trimToEmpty(v);
+        String val = v;
+        if (val != null) {
+            val = StringUtils.trimToEmpty(val);
         }
-
-        return v;
+        return val;
     }
 
     @Override
