+++ addons/binding/org.openhab.binding.denonmarantz/src/main/java/org/openhab/binding/denonmarantz/internal/connector/http/entities/types/OnOffType.java
@@ -1,18 +1,18 @@
 /**
- * Copyright (c) 2010-2016 by the respective copyright holders.
+ * Copyright (c) 2010-2018 by the respective copyright holders.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.denonmarantz.internal.communication.entities.types;
+package org.openhab.binding.denonmarantz.internal.connector.http.entities.types;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 
-import org.openhab.binding.denonmarantz.internal.communication.adapters.OnOffAdapter;
+import org.openhab.binding.denonmarantz.internal.connector.http.adapters.OnOffAdapter;
 
 /**
  * Contains an On/Off value in the form of a boolean
