+++ addons/binding/org.openhab.binding.denonmarantz/src/main/java/org/openhab/binding/denonmarantz/internal/connector/telnet/DenonMarantzTelnetListener.java
@@ -1,14 +1,14 @@
 /**
- * Copyright (c) 2010-2017 by the respective copyright holders.
+ * Copyright (c) 2010-2018 by the respective copyright holders.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.denonmarantz.internal.telnet;
+package org.openhab.binding.denonmarantz.internal.connector.telnet;
 
-import org.openhab.binding.denonmarantz.internal.DenonMarantzConnector;
+import org.openhab.binding.denonmarantz.internal.connector.DenonMarantzConnector;
 
 /**
  * Listener interface used to notify the {@link DenonMarantzConnector} about received messages over Telnet
